# kaifa-be/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kaifa API
  version: 1.0.0
  description: REST and WebSocket API for the Kaifa AI language learning platform.
servers:
  - url: https://api.kaifanesia.com/api
    description: Production API
  - url: http://localhost:8080/api
    description: Local API
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and issue access and refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /assessment/chat:
    post:
      tags: [Assessment]
      summary: Send one text turn to an assessment session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentChatRequest"
      responses:
        "200":
          description: Assessment reply or completed result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AssessmentChatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /assessment/speech/tts:
    post:
      tags: [Assessment]
      summary: Convert assessment text to speech
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeechRequest"
      responses:
        "200":
          description: MP3 audio stream
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /assessment/speech/stt:
    post:
      tags: [Assessment]
      summary: Transcribe assessment audio to text
      description: Accepts a complete audio file as a raw request body or multipart form field named audio.
      parameters:
        - in: query
          name: module_id
          schema:
            type: string
          required: false
          description: Module ID used to select the transcription language.
      requestBody:
        required: true
        content:
          audio/webm:
            schema:
              type: string
              format: binary
          audio/wav:
            schema:
              type: string
              format: binary
          audio/mpeg:
            schema:
              type: string
              format: binary
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
      responses:
        "200":
          description: Transcribed text
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TranscriptResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /learning/generate:
    post:
      tags: [Learning]
      summary: Generate a student-specific learning path after assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateLearningPathRequest"
      responses:
        "200":
          description: Generated learning path
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LearningPath"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /learning/chat:
    post:
      tags: [Learning]
      summary: Send one text turn to the AI tutor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningChatRequest"
      responses:
        "200":
          description: Tutor reply
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LearningChatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /learning/speech/stt:
    post:
      tags: [Learning]
      summary: Transcribe learning audio to text
      description: Accepts a complete audio file as a raw request body or multipart form field named audio.
      parameters:
        - in: query
          name: module_id
          schema:
            type: string
          required: false
          description: Module ID used to select the transcription language.
      requestBody:
        required: true
        content:
          audio/webm:
            schema:
              type: string
              format: binary
          audio/wav:
            schema:
              type: string
              format: binary
          audio/mpeg:
            schema:
              type: string
              format: binary
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
      responses:
        "200":
          description: Transcribed text
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TranscriptResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
    ErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: invalid request body
    CEFRLevel:
      type: string
      enum: [A1, A2, B1, B2, C1, C2]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: umar@kaifa.com
        password:
          type: string
          format: password
          example: password
    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - access_token_expires_in
        - refresh_token_expires_in
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        access_token_expires_in:
          type: integer
          format: int64
          example: 900
        refresh_token_expires_in:
          type: integer
          format: int64
          example: 604800
...
    AssessmentChatRequest:
      type: object
      required: [module_id, user_text]
      properties:
        module_id:
          type: string
          example: mod-eng-ix-01-01
        user_text:
          type: string
          example: My name is Ahmad. Nice to meet you.
    AssessmentChatResponse:
      type: object
      properties:
        reply:
          type: string
          example: Nice to meet you. Can you introduce a friend now?
        result:
          $ref: "#/components/schemas/AssessmentResult"
    AssessmentResult:
      type: object
      required: [cefr_level, level, score, title, summary, strengths, areas, translations]
      properties:
        cefr_level:
          $ref: "#/components/schemas/CEFRLevel"
        level:
          $ref: "#/components/schemas/CEFRLevel"
        score:
          type: integer
          minimum: 0
          maximum: 100
          example: 82
        title:
          type: string
          example: MasyaAllah, You Are Awesome
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        areas:
          type: array
          items:
            type: string
        translations:
          type: array
          items:
            type: object
            required: [language, title, summary, strengths, areas]
            properties:
              language:
                type: string
                example: bahasa indonesia
              title:
                type: string
              summary:
                type: string
              strengths:
                type: array
                items:
                  type: string
              areas:
                type: array
                items:
                  type: string
    SpeechRequest:
      type: object
      required: [user_text]
      properties:
        user_text:
          type: string
          example: Hello, welcome to today's lesson.
    TranscriptResponse:
      type: object
      required: [text]
      properties:
        text:
          type: string
          example: Hello, welcome to today's lesson.
    GenerateLearningPathRequest:
      type: object
      required: [module_id]
      properties:
        module_id:
          type: string
          example: mod-eng-ix-01-01
    LearningChatRequest:
      type: object
      required: [learning_path_step_id, user_text]
      properties:
        learning_path_step_id:
          type: string
          example: lp-7f9a1c2d3e4b5a60-step-1
        user_text:
          type: string
          example: I want to practice introducing myself.
...
    LearningChatResponse:
      type: object
      required: [type, reply]
      properties:
        type:
          type: string
          enum: [chat, guardrail, listening_paragraph, listening_question, listening_assessment, material_required]
          example: chat
        reply:
          type: string
        paragraph:
          type: string
          description: Present when type is listening_paragraph. The generated listening passage should be 100-150 words. Frontend can send this text to TTS.
        question_number:
          type: integer
          description: Present when type is listening_question.
          example: 1
        assessment:
          $ref: "#/components/schemas/ListeningAssessment"
        listening_state:
          $ref: "#/components/schemas/LearningListeningState"
    LearningListeningState:
      type: object
      properties:
        phase:
          type: string
          enum: [awaiting_ready, asking, completed]
          example: asking
        material_type:
          type: string
          enum: [listening, reading]
          example: listening
        paragraph:
          type: string
        questions_asked:
          type: integer
          example: 2
        answers:
          type: array
          items:
            type: string
    ListeningAssessment:
      type: object
      properties:
        score:
          type: integer
          example: 82
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        areas:
          type: array
          items:
            type: string
    LearningPath:
      type: object
      required: [id, language, module_id, level, topic_scope, steps, total_steps]
      properties:
        id:
          type: string
          example: lp-7f9a1c2d3e4b5a60
        language:
          type: string
          example: english
        module_id:
          type: string
          example: mod-eng-ix-01-01
        level:
          $ref: "#/components/schemas/CEFRLevel"
        topic_scope:
          type: string
        topic_scope_romanized:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/LearningStep"
        total_steps:
          type: integer
          example: 3
        translations:
          type: array
          items:
            $ref: "#/components/schemas/LearningPathTranslation"
    LearningStep:
      type: object
      required: [id, order, activity, title, description, required_time, topic_scope, target, progress]
      properties:
        id:
          type: string
          example: lp-7f9a1c2d3e4b5a60-step-1
        order:
          type: integer
          example: 1
        activity:
          type: string
          example: speaking
        activity_romanized:
          type: string
        title:
          type: string
        title_romanized:
          type: string
        description:
          type: string
        description_romanized:
          type: string
        required_time:
          type: integer
          description: Estimated time in minutes.
          example: 10
        topic_scope:
          type: string
          example: greetings and self-introduction
        topic_scope_romanized:
          type: string
        target:
          type: array
          items:
            type: string
          example: [simple present tense, subject pronouns]
        progress:
          type: integer
          minimum: 0
          maximum: 100
          example: 0
        translations:
          type: array
          items:
            $ref: "#/components/schemas/LearningStepTranslation"
    LearningPathTranslation:
      type: object
      required: [id, learning_path_id, language, topic_scope]
      properties:
        id:
          type: string
          example: lp-7f9a1c2d3e4b5a60-bi
        learning_path_id:
          type: string
          example: lp-7f9a1c2d3e4b5a60
        language:
          type: string
          example: bahasa indonesia
        topic_scope:
          type: string
    LearningStepTranslation:
      type: object
      required: [id, learning_path_step_id, language, activity, title, description, topic_scope]
      properties:
        id:
          type: string
          example: lp-7f9a1c2d3e4b5a60-step-1-bi
        learning_path_step_id:
          type: string
          example: lp-7f9a1c2d3e4b5a60-step-1
        language:
          type: string
          example: bahasa indonesia
        activity:
          type: string
        title:
          type: string
        description:
          type: string
        topic_scope:
          type: string
...